added additional tests, updated DEFINITION.md to reflect newly added file (test.sh)
