+++ DEFINTION.md
@@ -2,11 +2,13 @@ cerp/: This contains the python application (minus app.py and config.py).
 
 docs/: This contains the original pdf's with the data used within the entire application
 
-tests/: This (will) contain the python testing suit (unittest).
+tests/: This contains the python testing suit (unittest).
 
 vue_builder/: This contains the "front end" of the application.  Using vuejs
 
 .gitignore: This file tells git which other files should be ignored (build files, compiled files, cached libraries, etc)
+
+.travis.yml: This file tells travis-ci.org how to build and test our project.  It needs to be improved to include vue_builder tests
 
 app.py:  This is the main entry point for the application.  However while it's the entry point, it itself does nothing.
 
@@ -23,3 +25,5 @@ README.md: This file describes the project
 requirements.txt: this file lists all the python librarires that are used in this application.  It is auto generated using `pip freeze > requirements.txt`, if you add a new library, please re-execute that command to update the requirements file.  Heroku (our server) also uses this file so keeping it up to date is critial.
 
 start.sh: Starts the python server using linux commands.  Generally speaking, it exports two environ variables, then runs flask, which uses those variables to start the application
+
+test.sh: Runs the python tests
